fix(jdbc-dsl): propagate generic V through remaining type-safety gaps
- SqlFunctions: max/min preserve column type V; sum adds V extends Number;
  lag/lead/firstValue/lastValue preserve V instead of returning Object
- UpdateBuilder.set: Object→V to match column type
- WhereBuilder.inSubquery/notInSubquery: add SFunction overloads for
  ergonomic use alongside SqlExpression overloads
- Tests: align to new typed API (mapTo(Long.class), inSubquery, min
  instead of avg for integer-preserving scalar subquery tests,
  BigDecimal.valueOf for sum HAVING predicate)

All 364 tests pass.

+++ jdbc-dsl/src/test/java/io/github/jsbxyyx/jdbcdsl/SqlFunctionsTest.java
@@ -7,6 +7,8 @@ import io.github.jsbxyyx.jdbcdsl.expr.AggregateExpression;
 import io.github.jsbxyyx.jdbcdsl.expr.FunctionExpression;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 import static io.github.jsbxyyx.jdbcdsl.SqlFunctions.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -193,12 +195,12 @@ class SqlFunctionsTest {
         SelectSpec<TOrder, UserDto> spec = SelectBuilder.from(TOrder.class)
                 .select(col(TOrder::getStatus), sum(TOrder::getAmount))
                 .groupBy(TOrder::getStatus)
-                .having(h -> h.gte(sum(TOrder::getAmount), 1000))
+                .having(h -> h.gte(sum(TOrder::getAmount), BigDecimal.valueOf(1000)))
                 .mapTo(UserDto.class);
 
         RenderedSql rendered = SqlRenderer.renderSelect(spec);
         assertThat(rendered.getSql()).contains("HAVING SUM(t.amount) >= :p1");
-        assertThat(rendered.getParams()).containsEntry("p1", 1000);
+        assertThat(rendered.getParams()).containsEntry("p1", BigDecimal.valueOf(1000));
     }
 
     @Test
